Larger-refactor of how we track registers and instructions
This change includes a few changes to how we load ground-truth
information into gtirb-ddisasm-retypd, which we're moving towards
using the GTIRB as the "latest source of truth" rather than any
internal table, out of caution of certain APIs not being stable yet
and assuming that GTIRB will always have the "latest" information.
This is done for both block-level information and *some* instruction
level-information. For example, block and instruction boundaries are
now directly inferred from the GTIRB structure. This means that
gtirb-capstone is now a dependency of gtirb-ddisasm-retypd, and
consequentially we are re-decoding all instructions in the CFG.
We do now, however, have more control over information we can pull
from capstone, so this change-set also includes the beginnings of
a move away from just raw instruction operand information to access
information generated from capstone semantics. One consequence of
this change is now we no longer have to develop hacky ways to match
compares and sign-revelaing jumps, instead we can just track the
def-use chains of the flags from the compare to the jump. We do
also now have our own register tables. This is because the information
we want to track is not limited to just registers that might be
involved in computing addresses, we need to be able to track both
implicit and explicit register information, so now we can do so through
the new track_register table. This also includes register-alias
information so we could if it becomes useful stack tracking which
pieces of which register are aliasing eachother. The final big
change that comes with this is we are also now computing our own
def-use information, and doing so in an abstracted way such that
the def-use computation can be re-used in other def-use analyses
without having to re-implement the analysis. This is done in
def_use.dl.